src/main/java/com/falcon/tasks/service/impl/TaskServiceImpl.java
package com.falcon.tasks.service.impl;

import com.falcon.tasks.domain.Task;
import com.falcon.tasks.repository.TaskRepository;
import com.falcon.tasks.service.TaskService;
import org.springframework.stereotype.Service;

@Service
public class TaskServiceImpl implements TaskService {

    private final TaskRepository taskRepository;

    public TaskServiceImpl(TaskRepository taskRepository) {
        this.taskRepository = taskRepository;
    }

    @Override
    public Iterable<Task> getTasks() {
        return taskRepository.findAll();
    }

    @Override
    public Task save(Task task) {
        return taskRepository.save(task);
    }
}
